feat(web): auth boundary 인증 여부에 따른 분기 처리 구현
