--- apps/web/src/features/ai-recap/ui/AiRecapContent.tsx
@@ -1,51 +1,0 @@
-"use client";
-
-import type { RecapData } from "@recap/api";
-import { useSuspenseQuery } from "@tanstack/react-query";
-
-import { AuthConsumer } from "@/entities/auth/ui";
-import { aiRecapQueryOptions } from "@/features/ai-recap/api/ai-recap-query.client";
-import { hasRecapContent } from "@/features/ai-recap/lib/recap-mapper";
-import AiTimeline from "@/features/ai-recap/ui/AiTimeline";
-import RecapSummary from "@/features/ai-recap/ui/RecapSummary";
-import TopVisitedTopics from "@/features/ai-recap/ui/TopVisitedTopics";
-import AiRecapEmptyPage from "@/pages/ai-recap/ui/AiRecapEmptyPage";
-import AiRecapLoadingPage from "@/pages/ai-recap/ui/AiRecapLoadingPage";
-import AiRecapUnloginPage from "@/pages/ai-recap/ui/AiRecapUnloginPage";
-
-const AiRecapContent = ({ date }: { date: string }) => (
-  <AuthConsumer>
-    {({ isReady, isLoggedIn }) => {
-      if (!isReady) return <AiRecapLoadingPage />;
-      if (!isLoggedIn) return <AiRecapUnloginPage />;
-      return <LoggedInRecap date={date} />;
-    }}
-  </AuthConsumer>
-);
-
-const LoggedInRecap = ({ date }: { date: string }) => {
-  const {
-    data: recap,
-    isLoading,
-    isError,
-  } = useSuspenseQuery({
-    ...aiRecapQueryOptions(date),
-    select: (data): RecapData | null => {
-      const recap = data.data;
-      return recap && hasRecapContent(recap) ? recap : null;
-    },
-  });
-
-  if (isLoading) return <AiRecapLoadingPage />;
-  if (isError || !recap) return <AiRecapEmptyPage />;
-
-  return (
-    <>
-      <RecapSummary recap={recap} />
-      <AiTimeline timelines={recap.timelines ?? []} />
-      <TopVisitedTopics topics={recap.topics ?? []} />
-    </>
-  );
-};
-
-export default AiRecapContent;
